requirements.txt
grpcio-tools==1.43.0

protobuf==3.17.3